requirements/base.txt
django>=5.1.0
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=2.30.0
python-dotenv>=1.0.0
langdetect>=1.0.9
